# app-engine/src/test/resources/spec/api/openapi_spec.v0.1.0.yml
openapi: 3.0.0
info:
  title: App Engine API
  version: 0.2.0
  description: App Engine API provides a set of requests to interact with tasks and apps.
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8080/api/v1
paths:
  /tasks:
    post:
      summary: Upload and creates a task as a zip archive containing a 'descriptor.yml' file and related task files
      operationId: uploadTask
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                task:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task_description"
        '400':
          description: The descriptor is malformed, e.g. not valid based on the schema, missing files or contains erroneous information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
        '409':
          description: Conflict, a task with the same namespace and version has already been registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'

    get:
      summary: Get descriptions of all tasks available in App Engine
      operationId: getTasks
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/task_description'
  /tasks/{id}:
    get:
      summary: Get task description by UUID
      operationId: getTaskByUUID
      parameters: &task-uuid-only-parameters
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task
      responses: &get-task-description-responses
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task_description'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{namespace}/{version}:
    get:
      summary: Get task description by namespace and version
      operationId: getTaskByNamespaceVersion
      parameters: &task-ns-version-only-parameters
        - in: path
          name: namespace
          required: true
          schema:
            type: string
          description: The namespace of the task
        - in: path
          name: version
          required: true
          schema:
            type: string
          description: The version of the task
      responses: *get-task-description-responses
...
  /tasks/{namespace}/{version}/runs:
    post:
      summary: creates a Task run
      operationId: createTaskRunByNamespaceVersion
      parameters: *task-ns-version-only-parameters
      responses: &post-task-run-responses
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task_run'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /tasks/{task_id}/runs:
    post:
      summary: creates a Task run
      operationId: createTaskRunByUUID
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task
      responses: *post-task-run-responses
...
  /task-runs/{run_id}:
    get:
      summary: Get information about a task run
      operationId: getTaskRun
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the task run
      responses:
        '200':
          description: returns the task run description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task_run"
        '404':
          description: task run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    contact:
      type: object
      properties:
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        organization:
          type: string
          description: This person's organization
        email:
          type: string
          description: Email address
        is_contact:
          type: boolean
          description: This person is the contact person
...
    task_description:
      type: object
      required:
        - name
        - namespace
        - version
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Task name
        namespace:
          type: string
          description: Task namespace
        version:
          type: string
          description: Task version
        description:
          type: string
          description: Task description
        authors:
          type: array
          description: List of the task authors
          items:
            $ref: "#/components/schemas/contact"
...
    task_run_state:
      type: string
      enum:
        - CREATED      # created but not all inputs provisioned
        - PROVISIONED  # ready to be executed, all inputs have been provisioned
        - QUEUING      # submitting the task to the execution environment
        - QUEUED       # evaluated successfully by execution environment
        - PENDING      # pending execution on the execution environment
        - RUNNING      # task running in the execution environment
        - FAILED       # an error occurred and stopped the process of executing the task (terminal state)
        - FINISHED     # task successfully executed and outputs available (terminal state)
    task_run:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/task_description"
        id:
          type: string
          format: uuid
          description: Parameter App Engine task run UUID
        state:
          $ref: "#/components/schemas/task_run_state"
        created_at:
          description: Creation datetime for the task run
          type: string
          format: date-time
        updated_at:
          description: Last update datetime for the task run
          type: string
          format: date-time
...
    generic_error:
      type: object
      properties:
        errorCode:
          type: string
          pattern: ^APPE-[a-zA-Z0-9_-]+$
          description: an error code structured as `{component}-{reason}`
        message:
          type: string
          description: the human-readable error message associated with the errorCode (non-localized, in english)
        details:
          type: object
          description: provide a more detailed description of the error, structure and content depends on the error code
...